refactor: updates authentication middleware name in files
