+++ src/routes/ordersRouter.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import validateToken from "../middlewares/validateTokenMiddleware.js";
+import validateToken from "../middlewares/authenticationMiddleware.js";
 import validateOrder from "../middlewares/validateOrderMiddleware.js";
 import createNewOrder from "../controllers/orderController.js";
